requirements.txt
# Web framework
flask>=3.1.2

# Microsoft Authentication Library
msal>=1.34.0

# HTTP client
requests>=2.32.3

# Environment variable loader
python-dotenv>=1.0.0

# MongoDB driver (PyMongo)
pymongo>=4.6.0

# HTML to Markdown conversion (used when processing email bodies)
html2text>=2020.1.16